server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bounded-pool-executor==0.0.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.5
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
pillow==11.0.0
ping3==4.0.8
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.19
python-pptx==1.0.2
python-slugify==8.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
starlette==0.41.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
XlsxWriter==3.2.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.1
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
crewai==0.5.0
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.5.14
datasets==3.1.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastai==2.7.18
fastapi==0.113.0
fastapi-cli==0.0.5
fastbook==0.0.29
fastcore==1.7.23
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.14.0
filetype==1.2.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
googleapis-common-protos==1.63.2
graphviz==0.20.3
groq==0.13.1
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.8.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.11
kiwisolver==1.4.5
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langchain-text-splitters==0.2.2
langcodes==3.5.0
langsmith==0.0.87
language_data==1.3.0
Lazify==0.4.0
literalai==0.0.607
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.11
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.37.1
opencv-python==4.9.0.80
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prometheus_client==0.21.1
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyppeteer==2.0.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.22.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tenacity==8.5.0
terminado==0.18.1
thinc==8.3.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.5.2
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
uptrace==1.26.0
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==0.20.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
widgetsnbextension==3.6.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.18.2
